pi-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.9.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.9
httpx==0.27.0
psutil==5.9.8
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
